+++ src/follow.c
@@ -4,10 +4,10 @@
 #include <dsp/follow.h>
 #include <dsp/utils.h>
 
-static const float TLD_AUDIO_ENVELOPE_ANALOG_TC = -0.99967234081320612357829304641019;
+static const float TLD_AUDIO_ENVELOPE_ANALOG_TC = -0.99967234081320612357829304641019f;
 
 static inline float calculate_time_(float t_ms, float sr) {
-    return exp(TLD_AUDIO_ENVELOPE_ANALOG_TC / (t_ms * sr * 0.000001));
+    return exp(TLD_AUDIO_ENVELOPE_ANALOG_TC / (t_ms * sr * 0.000001f));
 }
 
 static inline void update_atk_time_(follow* self) {
@@ -19,7 +19,7 @@ static inline void update_rel_time_(follow* self) {
 }
 
 static inline float peak_tick_(follow* self, float xn) {
-    float current_envelope = 0.0;
+    float current_envelope = 0.0f;
     float last_envelope = self->last_envelope_;
     xn = fabs(xn);
 
@@ -29,14 +29,14 @@ static inline float peak_tick_(follow* self, float xn) {
         current_envelope = self->rel_time_ * (last_envelope - xn) + xn;
     }
     current_envelope = check_float_underflow(current_envelope);
-    current_envelope = dsp_max(current_envelope, 0.0);
+    current_envelope = dsp_max(current_envelope, 0.0f);
 
     self->last_envelope_ = current_envelope;
     return current_envelope;
 }
 
 static inline float rms_tick_(follow* self, float xn) {
-    float current_envelope = 0.0;
+    float current_envelope = 0.0f;
     float last_envelope = self->last_envelope_;
 
     xn = fabs(xn);
@@ -49,9 +49,9 @@ static inline float rms_tick_(follow* self, float xn) {
     }
 
     current_envelope = check_float_underflow(current_envelope);
-    current_envelope = dsp_max(current_envelope, 0.0);
+    current_envelope = dsp_max(current_envelope, 0.0f);
 
-    current_envelope = pow(current_envelope, 0.5);  // sqrt for rms;
+    current_envelope = powf(current_envelope, 0.5f);  // sqrt for rms;
     self->last_envelope_ = current_envelope;
     return current_envelope;
 }
@@ -60,7 +60,7 @@ void follow_init(follow* self, float atk_ms, float rel_ms, float sr) {
     self->atk_ms = atk_ms;
     self->rel_ms = rel_ms;
     self->sr = sr;
-    self->last_envelope_ = 1e-9;
+    self->last_envelope_ = 1e-9f;
 
     update_atk_time_(self);
     update_rel_time_(self);
